+++ examples/complete/main.tf
@@ -26,11 +26,7 @@ module "ecs_service_alb" {
     subnets          = local.private_subnets
     assign_public_ip = true
   }
-
-  triggers = {
-    redeployment = timestamp()
-  }
-
+  # Always redeploy the service when terraform apply is run
   alb_subnets                       = local.public_subnets
   cluster                           = local.cluster
   vpc_id                            = local.vpc_id
@@ -46,10 +42,10 @@ module "ecs_service_alb" {
   tasks_maximum_percent             = 150
   tags                              = local.tags
 
-  load_balancer = {
+  load_balancer = [{
     container_name = var.name
     container_port = var.containerport
-  }
+  }]
 
   access_logs = {
     bucket  = module.access_logs_bucket.id
@@ -61,10 +57,49 @@ module "ecs_service_alb" {
   drop_invalid_header_fields = var.drop_invalid_header_fields
   tg_port                    = var.tg_port
   create_load_balancer       = var.create_load_balancer
-  enable_autoscaling         = var.enable_autoscaling
-  scalable_dimension         = var.scalable_dimension
-  service_namespace          = var.service_namespace
-  metric_aggregation_type    = "Average"
+  enable_autoscaling         = true
+  step_scaling_policies = [
+    {
+      name        = "cpu-policy"
+      policy_type = "StepScaling"
+      step_scaling_policy_configuration = {
+        adjustment_type         = "ChangeInCapacity"
+        cooldown                = 300
+        metric_aggregation_type = "Average"
+        step_adjustments = [
+          {
+            metric_interval_lower_bound = 0
+            metric_interval_upper_bound = 10
+            scaling_adjustment          = 1
+          },
+          {
+            metric_interval_lower_bound = 10
+            metric_interval_upper_bound = null
+            scaling_adjustment          = 2
+          }
+        ]
+      }
+    }
+  ]
+  scheduled_actions = [
+    {
+      name         = "ecs-scheduled-scale-out"
+      schedule     = "cron(0 8 ? * MON-FRI *)"
+      min_capacity = 5
+      max_capacity = 10
+      timezone     = "UTC"
+    },
+    {
+      name         = "ecs-scheduled-scale-in"
+      schedule     = "cron(0 20 ? * MON-FRI *)"
+      min_capacity = 2
+      max_capacity = 4
+      timezone     = "UTC"
+    }
+  ]
+  scalable_dimension = var.scalable_dimension
+  service_namespace  = var.service_namespace
+  # metric_aggregation_type    = "Average"
 
   # WAF association
   associate_with_waf = true
@@ -89,6 +124,7 @@ module "ecs_service_nlb" {
   family                   = "${var.name}-nlb-task-definition"
   enable_execute_command   = var.enable_execute_command
   load_balancer_type       = "network"
+  # Always redeploy the service when terraform apply is run
   network_configuration = {
     subnets          = local.private_subnets
     assign_public_ip = true
@@ -108,10 +144,10 @@ module "ecs_service_nlb" {
   tasks_maximum_percent             = 150
   tags                              = local.tags
 
-  load_balancer = {
+  load_balancer = [{
     container_name = var.name
     container_port = var.containerport
-  }
+  }]
   access_logs = {
     bucket  = module.access_logs_bucket.id
     enabled = var.access_logs_enabled
@@ -123,9 +159,36 @@ module "ecs_service_nlb" {
   tg_port                    = var.tg_port
   tg_protocol                = "TCP"
   create_load_balancer       = var.create_load_balancer
-  enable_autoscaling         = var.enable_autoscaling
-  scalable_dimension         = var.scalable_dimension
-  service_namespace          = var.service_namespace
+  enable_autoscaling         = true
+  target_scaling_policies = [
+    {
+      name        = "cpu-target-tracking"
+      policy_type = "TargetTrackingScaling"
+      target_tracking_scaling_policy_configuration = {
+        target_value       = 75
+        scale_in_cooldown  = 300
+        scale_out_cooldown = 300
+        predefined_metric_specification = {
+          predefined_metric_type = "ECSServiceAverageCPUUtilization"
+        }
+      }
+    },
+    {
+      name        = "memory-target-tracking"
+      policy_type = "TargetTrackingScaling"
+      target_tracking_scaling_policy_configuration = {
+        target_value       = 80
+        scale_in_cooldown  = 300
+        scale_out_cooldown = 300
+        predefined_metric_specification = {
+          predefined_metric_type = "ECSServiceAverageMemoryUtilization"
+        }
+      }
+    }
+  ]
+
+  scalable_dimension = var.scalable_dimension
+  service_namespace  = var.service_namespace
 
   # Load balancer sg
   lb_ingress_rules = var.nlb_ingress_rules