+++ CHANGELOG.md
@@ -26,6 +26,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - feat: consolidate ecs cluster module and ecs service module into one
 - feat: Add missing aws_ecs_task_definition arguments and showcase them in examples
 - feat: expand volume block of the task definition as it has more configuration
+- feat: Add example for service security group using a security group id for `service_ingress_sg`
+
+## [1.11.0] - 2024-05-28
+### Changes
+- feat: Add Step Scaling Policy to ECS Service
+- feat: Add Target Tracking Policy to ECS Service
+- feat: Add Scheduled Scaling Policy to ECS Service
+- feat: Add Application Auto Scaling examples
+- fix: change the trigger to plaintimestamp on examples to fix "registry.terraform.io/hashicorp/aws" produced an invalid new value for .triggers["redeployment"]
+- tflint: Change the loab balancer block to use a list of maps instead of a map of maps, **this is a breaking change.**
+- tflint: Change the reference of index for ecs task definition roles.
+- docs: Add better description to some variables.
 
 ## [1.10.0] - 2024-05-19
 ### Changes
@@ -156,8 +168,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - feat: feature update.
 - feat: initial code commit
 
-[Unreleased]: https://github.com/boldlink/terraform-aws-ecs-service/compare/1.10.0...HEAD
+[Unreleased]: https://github.com/boldlink/terraform-aws-ecs-service/compare/1.11.0...HEAD
 
+[1.11.0]: https://github.com/boldlink/terraform-aws-ecs-service/releases/tag/1.11.0
 [1.10.0]: https://github.com/boldlink/terraform-aws-ecs-service/releases/tag/1.10.0
 [1.9.0]: https://github.com/boldlink/terraform-aws-ecs-service/releases/tag/1.9.0
 [1.8.0]: https://github.com/boldlink/terraform-aws-ecs-service/releases/tag/1.8.0